+++ front/public/main.js
@@ -5,7 +5,7 @@ import './components/icons/ui-icon.ce.js';
 import './views/view-signup.ce.js';
 import './views/view-profil.ce.js';
 import './views/view-email-confirmation.ce.js';
-import './views/login.js';
+import './views/view-login.ce.js';
 import './game/view-game.ce.js';
 
 // router
